Version 0.2b2

+++ websvn-admin/index.php
@@ -1,5 +1,3 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="ja" lang="ja">
 <?php
 
 $strVersion = '0.2';	// 画面に表示するバージョン
@@ -31,42 +29,33 @@ $strVersion = '0.2';	// 画面に表示するバージョン
 // http://www.opensource.jp/gpl/gpl.ja.html
 // ******************************************************
 
-?>
-<head>
-	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-	<meta http-equiv="Content-Language" content="ja" />
-	<link rel="stylesheet" href="style.css" type="text/css" />
-
-	<title> </title>
-
-	<script type="text/javascript" src="../utf.js"></script>
-	<script type="text/javascript" src="../md5.js"></script>
-	<script type="text/javascript" src="../authpage_form_md5.js"></script>
-
-</head>
-<body>
-
-<div style="height:100px; width:100%; padding:0px; margin:0px;">
-<p><img src="./logo-svn.png" width="109" height="93" alt="Subversion" style="vertical-align:middle;" /><span style="margin:0px 20px; font-size:30px; font-weight:lighter;">WebSVN-Admin</span><span style="margin:0px 0px; font-size:25px; font-weight:lighter; color:lightgray;">Subversion Administration</span></p>
-</div>
-
-<?php
-
 require_once('include/config.php');	// ディレクトリなどの設定
 require_once('include/auth.php');		// ユーザ認証
 
 // このスクリプトのファイル名
 $strFilenameThis = htmlspecialchars(basename($_SERVER['PHP_SELF']));
 
-// config.php で設定が行われているか確認する
-if(!isset($strSvnCmdPath) || !isset($strBaseDir) || !isset($strBackupDir)){
-	print("<p class=\"error\">include/config.php の初期設定が行われていません</p>\n");
-	print("</body>\n</html>\n");
+// 認証ファイルの確認
+// auth.datの存在確認 1:OK, 0:Create, -1:Create Error
+$flag_auth_file_exist = CheckAuthDataFile();
+
+// ファイルのダウンロード（ヘッダを表示する前に行い、終了する）
+if(isset($_GET['mode']) && $_GET['mode'] === 'download' && isset($_GET['dumptext'])){
+	$strRepo = $_GET['dumptext'];
+	download_dump_text($strRepo);
+}
+
+// HTML構文を開始し、ヘッダー部分を表示する
+print_html_header();
+
+// include/config.php 内のグローバル変数が設定されているかチェック
+if(!check_config_file(1)){
+	print_html_footer();
 	die;
 }
 
-$nResult = CheckAuthDataFile();
-if($nResult == 0){
+// 認証ファイルの確認結果を表示（必要な場合）
+if($flag_auth_file_exist == 0){
 ?>
 	<p class="info">初期ユーザ名：user, パスワード：password です</p>
 	<p><a href="<?php echo $strFilenameThis;?>">ログオン画面を表示する</a><p>
@@ -75,16 +64,14 @@ if($nResult == 0){
 <?php
 	die;
 }
-elseif($nResult < 0){
+elseif($flag_auth_file_exist < 0){
 	print("<p class=\"error\">認証用データファイルが作成できません<br />dataディレクトリに書き込み権限が無い可能性があります</p>\n");
-	print("</body>\n</html>\n");
+	print_html_footer();
 	die;
 }
 
 // ユーザ認証を行う
-$strReturn = CheckAuth($strFilenameThis, 'svnadmin-create');
-
-if(!$strReturn)
+if(!CheckAuth($strFilenameThis, 0))
 {
 	print("<p>認証が行われていません。またはCookieが使えない状況です。</p>\n");
 	print("<a href=\"".$strFilenameThis."\">再度ログオン画面を表示する</a>\n");
@@ -97,7 +84,6 @@ if(!$strReturn)
 <div id="main_content_left">
 <h2>System</h2>
 <p><?php echo date('Y/m/d  H:i:s', time()) ; ?><br />
-WebSVN-Admin &nbsp;<?php echo $strVersion; ?><br />
 Subversion &nbsp;<?php echo get_svn_version(); ?></p>
 <h2>Menu</h2>
 <ul>
@@ -109,9 +95,7 @@ Subversion &nbsp;<?php echo get_svn_version(); ?></p>
 <h2>Repositories</h2>
 <?php
 
-// *********************
 // 既存リポジトリ一覧を表示（左側ペイン）
-// *********************
 display_repositories();
 
 ?>
@@ -121,89 +105,68 @@ display_repositories();
 <?php
 
 // *********************
-// バックアップディレクトリの一覧
+// プログラム引数によって処理分岐
 // *********************
+
+// バックアップディレクトリの一覧
 if(isset($_GET['mode']) && $_GET['mode'] === 'backuplist'){
 	display_backup_list();
 }
-// *********************
 // 新規リポジトリ作成
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'makerepo' && isset($_POST['newrepo']) && strlen($_POST['newrepo'])>0){
 	$strRepo = trim($_POST['newrepo']);
 	create_new_repository($strRepo);
 }
-// *********************
 // ログアウト
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'logout'){
 	LogoffAuth();
-?>
-<h1>Logout</h1>
-<p>ログアウトしました</p>
-<?php
+	print("<h1>Logout</h1>\n<p>ログアウトしました</p>\n");
 }
-// *********************
 // ユーザ名・パスワード変更
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'chgpasswd'){
 	print("<h1>Change User and Password (ユーザ名、パスワード変更)</h1>\n");
 	print("<p>".ChangePassword($strFilenameThis, 'svnadmin-create')."</p>\n");
 }
-// *********************
 // 既存リポジトリの情報表示（バックアップ、削除サブメニュー表示）
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'inforepo' && isset($_GET['reponame'])){
 	$strRepo = $_GET['reponame'];
 	info_repository($strRepo);
 }
-// *********************
 // 既存リポジトリのベリファイ
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'verify' && isset($_GET['reponame'])){
 	$strRepo = $_GET['reponame'];
 	verify_repository($strRepo);
 }
-// *********************
 // 既存リポジトリのバックアップ（hotcopy）
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'hotcopy' && isset($_GET['reponame'])){
 	$strRepo = $_GET['reponame'];
 	hotcopy_repository($strRepo, 0);
 }
-// *********************
 // 既存リポジトリのバックアップ（dump）
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'dump' && isset($_GET['reponame'])){
 	$strRepo = $_GET['reponame'];
 	dump_repository($strRepo);
 }
-// *********************
 // リポジトリの削除
-// *********************
 elseif(isset($_GET['mode']) && $_GET['mode'] === 'remove' && isset($_GET['reponame'])){
 	$strRepo = $_GET['reponame'];
 	remove_repository($strRepo);
 }
-// *********************
+// 既存リポジトリの情報表示（バックアップ、削除サブメニュー表示）
+elseif(isset($_GET['mode']) && $_GET['mode'] === 'viewbackupdir' && isset($_GET['dirname'])){
+	$strRepo = $_GET['dirname'];
+	view_backup_dir($strRepo);
+}
 // 新規リポジトリ作成 入力画面
-// *********************
 else{
 	// 引数が何もなかった場合、新規リポジトリ名の入力画面を表示
 	input_new_repository();
 }
 
-?>
-</div>	<!-- id="main_content_right" -->
-<p>&nbsp;</p>
-<div class="clear"></div>
-<div id="footer">
-<p><a href="http://sourceforge.jp/projects/websvn-admin/">WebSVN-Admin</a> version <?php echo $strVersion; ?> &nbsp;&nbsp; GNU GPL free software</p>
-</div>	<!-- id="footer" -->
+print('</div>	<!-- id="main_content_right" -->'."\n");
 
-</body>
-</html>
-<?php
+// HTMLのフッターを表示する(HTML構文を閉じる)
+print_html_footer();
 
 exit();
 
@@ -212,6 +175,7 @@ exit();
 // *********************
 function display_repositories() {
 	global $strBaseDir;
+	global $strFilenameThis;
 
 	$arrDirs = array();
 	if ($dir = opendir($strBaseDir)) {
@@ -235,9 +199,55 @@ function display_repositories() {
 
 
 // *********************
+// HTML構文を開始し、ヘッダー部分を表示する
+// *********************
+function print_html_header() {
+?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="ja" lang="ja">
+<head>
+	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+	<meta http-equiv="Content-Language" content="ja" />
+	<link rel="stylesheet" href="style.css" type="text/css" />
+
+	<title> </title>
+
+	<script type="text/javascript" src="../utf.js"></script>
+	<script type="text/javascript" src="../md5.js"></script>
+	<script type="text/javascript" src="../authpage_form_md5.js"></script>
+
+</head>
+<body>
+
+<div style="height:100px; width:100%; padding:0px; margin:0px;">
+<p><img src="./logo-svn.png" width="109" height="93" alt="Subversion" style="vertical-align:middle;" /><span style="margin:0px 20px; font-size:30px; font-weight:lighter;">WebSVN-Admin</span><span style="margin:0px 0px; font-size:25px; font-weight:lighter; color:lightgray;">Subversion Administration</span></p>
+</div>
+<?php
+}
+
+// *********************
+// フッター部分を表示した後、HTML構文を閉じる
+// *********************
+function print_html_footer() {
+	global $strVersion;
+?>
+<p>&nbsp;</p>
+<div class="clear"></div>
+<div id="footer">
+<p><a href="http://sourceforge.jp/projects/websvn-admin/">WebSVN-Admin</a> version <?php echo $strVersion; ?> &nbsp;&nbsp; GNU GPL free software</p>
+</div>	<!-- id="footer" -->
+
+</body>
+</html>
+<?php
+}
+
+
+// *********************
 // 新規リポジトリ名入力画面
 // *********************
 function input_new_repository() {
+	global $strFilenameThis;
 ?>
 
 <h1>Create New Repository (リポジトリ作成)</h1>
@@ -301,6 +311,7 @@ function create_new_repository($strRepo) {
 function info_repository($strRepo) {
 	global $strSvnCmdPath;
 	global $strBaseDir;
+	global $strFilenameThis;
 
 	print("<h1>Repository Administration (リポジトリ管理)</h1>\n");
 	print("<p>リポジトリ名 : ".htmlspecialchars($strRepo)."</p>\n");
@@ -316,7 +327,6 @@ function info_repository($strRepo) {
 	exec($strSvnCmdPath."svnlook youngest ".$strBaseDir.$strRepo." 2>&1", $arrStdout, $nResult);
 	if($nResult != 0){
 		print("<p class=\"error\">svnlook youngestコマンドが実行できません</p>\n");
-		return;
 	}
 	if(count($arrStdout) >= 1){ $strRevNo = $arrStdout[0]; }
 	$arrStdout = array();
@@ -326,7 +336,6 @@ function info_repository($strRepo) {
 	exec($strSvnCmdPath."svnlook author ".$strBaseDir.$strRepo." 2>&1", $arrStdout, $nResult);
 	if($nResult != 0){
 		print("<p class=\"error\">svnlook authorコマンドが実行できません</p>\n");
-		return;
 	}
 	if(count($arrStdout) >= 1){ $strAuthor = $arrStdout[0]; }
 	$arrStdout = array();
@@ -336,7 +345,6 @@ function info_repository($strRepo) {
 	exec($strSvnCmdPath."svnlook date ".$strBaseDir.$strRepo." 2>&1", $arrStdout, $nResult);
 	if($nResult != 0){
 		print("<p class=\"error\">svnlook dateコマンドが実行できません</p>\n");
-		return;
 	}
 	if(count($arrStdout) >= 1){ $strDate = $arrStdout[0]; }
 	
@@ -398,6 +406,8 @@ function verify_repository($strRepo) {
 // *********************
 // 既存リポジトリのバックアップ（hotcopy）
 // *********************
+// 引数 $flag_mode : 0のとき、処理のタイトルを表示する
+// 戻り値 0:失敗, 1:成功
 function hotcopy_repository($strRepo, $flag_mode) {
 	global $strSvnCmdPath;
 	global $strBaseDir;
@@ -411,14 +421,14 @@ function hotcopy_repository($strRepo, $flag_mode) {
 	// リポジトリ名に不正な文字が入っていないか検査
 	if(!preg_match("/^[A-Za-z0-9\-]+$/", $strRepo) || $strRepo[0] == '-' || $strRepo[strlen($strRepo)-1] == '-' || strlen($strRepo) > 20 || strlen($strRepo) <= 0){
 		print("<p class=\"error\">不正なリポジトリ名が指定されました</p>\n");
-		return;
+		return(0);
 	}
 
 	// バックアップ先ディレクトリが既存でないか検査
 	$strBackupBasename = $strRepo.'-'.date('Ymd-Hi', time());
 	if(file_exists($strBackupDir.$strBackupBasename)){
 		print("<p class=\"error\">バックアップ先に同じ名称のディレクトリがあります</p>\n");
-		return;
+		return(0);
 	}
 
 	print("<p>コマンド実行中 ... (svnadmin hotcopy ".htmlspecialchars($strRepo)." ".htmlspecialchars($strBackupBasename).")</p>\n");
@@ -426,8 +436,24 @@ function hotcopy_repository($strRepo, $flag_mode) {
 	// バックアップコマンドを実行
 	exec($strSvnCmdPath."svnadmin hotcopy ".$strBaseDir.$strRepo." ".$strBackupDir.$strBackupBasename." 2>&1", $arrStdout, $nResult);
 	if($nResult != 0){
-		print("<p class=\"error\">バックアップに失敗しました</p>\n");
-		return;
+		print("<p class=\"info\">バックアップに失敗しました。単純コピーを試します</p>\n");
+		print("<p>コマンド実行中 ... (cp -Rv ".htmlspecialchars($strRepo)." ".htmlspecialchars($strBackupBasename).")</p>\n");
+		$arrStdout = array();
+		exec("cp -Rv ".$strBaseDir.$strRepo." ".$strBackupDir.$strBackupBasename." 2>&1", $arrStdout, $nResult);
+		// コマンドのStdout出力がある場合
+		if(count($arrStdout)>0){
+			print("<pre>\n\n");
+			foreach($arrStdout as $str){
+				print($str."\n");
+			}
+			print("</pre>\n");
+		}
+		if($nResult != 0){
+			print("<p class=\"error\">単純コピーも出来ません。バックアップに失敗しました</p>\n");
+			return(0);
+		}
+		print("<p class=\"ok\">hotcopyは失敗しましたが、単純コピーは完了しました</p>\n");
+		return(1);
 	}
 	// コマンドのStdout出力がある場合
 	if(count($arrStdout)>0){
@@ -440,7 +466,7 @@ function hotcopy_repository($strRepo, $flag_mode) {
 	else{
 		print("<p class=\"ok\">バックアップが完了しました</p>\n");
 	}
-	return;
+	return(1);
 }
 
 // *********************
@@ -495,6 +521,7 @@ function dump_repository($strRepo) {
 function remove_repository($strRepo) {
 	global $strSvnCmdPath;
 	global $strBaseDir;
+	global $strBackupDir;
 
 	print("<h1>Remove Repository (リポジトリ削除)</h1>\n");
 	print("<p>リポジトリ名 : ".htmlspecialchars($strRepo)."</p>\n");
@@ -507,10 +534,12 @@ function remove_repository($strRepo) {
 
 	// バックアップ（hotcopy）
 	print("<p class=\"info\">削除前に、バックアップを行います</p>\n");
-	hotcopy_repository($strRepo, 1);
+	if(!hotcopy_repository($strRepo, 1)){
+		print("<p class=\"error\">対象リポジトリがバックアップできないため、削除を中止します</p>\n");
+		return;
+	}
 
 	print("<p>コマンド実行中 ... (rm -rfv ".htmlspecialchars($strRepo).")</p>\n");
-
 	// 削除
 	exec("rm -rfv ".$strBaseDir.$strRepo." 2>&1", $arrStdout, $nResult);
 	if($nResult != 0){
@@ -536,6 +565,7 @@ function remove_repository($strRepo) {
 // *********************
 function display_backup_list() {
 	global $strBackupDir;
+	global $strFilenameThis;
 
 	print("<h1>Backup Catalog</h1>\n");
 
@@ -553,10 +583,87 @@ function display_backup_list() {
 
 	print("<ul>\n");
 	foreach($arrDirs as $val){
-		print "<li>".htmlspecialchars($val)."</li>\n";
+		if(is_dir($strBackupDir.$val)){
+			print "<li><a href=\"".$strFilenameThis."?mode=viewbackupdir&amp;dirname=".htmlspecialchars($val)."\">".htmlspecialchars($val)."</a></li>\n";
+		}
+		else{
+			print "<li><a href=\"".$strFilenameThis."?mode=download&amp;dumptext=".htmlspecialchars($val)."\">".htmlspecialchars($val)."</a></li>\n";
+		}
 	}
 	print("</ul>\n");
 
+}
+
+
+// *********************
+// バックアップ一覧を表示する（ディレクトリの場合）
+// *********************
+function view_backup_dir($strRepo) {
+	global $strBackupDir;
+
+	print("<h1>View Backup Repository (hotcopy dir)</h1>\n");
+	print("<p>リポジトリ名 : ".htmlspecialchars($strRepo)."</p>\n");
+
+	// リポジトリ名に不正な文字が入っていないか検査
+	if(!preg_match("/^[A-Za-z0-9\-]+$/", $strRepo) || $strRepo[0] == '-' || $strRepo[strlen($strRepo)-1] == '-' || strlen($strRepo) > 40 || strlen($strRepo) <= 0){
+		print("<p class=\"error\">不正なリポジトリ名が指定されました</p>\n");
+		return;
+	}
+
+	print("<p>コマンド実行中 ... (ls -laR ".htmlspecialchars($strRepo).")</p>\n");
+	// ファイル一覧
+	exec("ls -laR ".$strBackupDir.$strRepo." 2>&1", $arrStdout, $nResult);
+	if($nResult != 0){
+		print("<p class=\"error\">一覧取得に失敗しました</p>\n");
+		return;
+	}
+	// コマンドのStdout出力がある場合
+	if(count($arrStdout)>0){
+		print("<pre>\n\n");
+		print("$ ls -laR ".htmlspecialchars($strRepo)."\n\n");
+		foreach($arrStdout as $str){
+			print($str."\n");
+		}
+		print("</pre>\n");
+	}
+
+}
+
+
+// *********************
+// dump(バックアップ)ファイルをダウンロードする
+// *********************
+function download_dump_text($strRepo) {
+	global $strFilenameThis;
+	global $strBackupDir;
+
+	// 認証確認
+	if(!CheckAuth($strFilenameThis, 1)){
+		die("Error : not logged on\n");
+	}
+
+	// include/config.php 内のグローバル変数が設定されているかチェック
+	if(!check_config_file(0)){
+		die("Error : config value\n");
+	}
+
+	// リポジトリ名に不正な文字が入っていないか検査
+	if(!preg_match("/^[A-Za-z0-9\-\.]+$/", $strRepo) || $strRepo[0] == '-' || $strRepo[strlen($strRepo)-1] == '-' || strlen($strRepo) > 40 || strlen($strRepo) <= 0){
+		die("Error : unsupported filename\n");
+	}
+
+	$strFilename = $strBackupDir.$strRepo;
+	if(!is_file($strFilename)){
+		die("file '$strFilename' not exist\n");
+	}
+
+	$file_length = filesize($strFilename);
+	header("Content-Disposition: attachment; filename=$strFilename");
+	header("Content-Length:$file_length");
+	header("Content-Type: application/octet-stream");
+	readfile ($strFilename);
+	
+	exit();
 }
 
 
@@ -577,6 +684,46 @@ function get_svn_version() {
 	}
 	return($strSvnVersion);
 }
+
+
+// *********************
+// config.php が正しく設定されているかチェック
+// *********************
+function check_config_file($flag_in_html) {
+	global $strSvnCmdPath;
+	global $strBaseDir;
+	global $strBackupDir;
+	
+	$flag_ok = 1;	// 正しくセットされていれば 1, 設定が足りなければ 0
+
+	if(!isset($flag_in_html)){ $flag_in_html = 0; }
+
+	// config.php で設定が行われているか確認する
+	$arrVarName = array(
+		array($strSvnCmdPath, '$strSvnCmdPath'),
+		array($strBaseDir, '$strBaseDir'),
+		array($strBackupDir, '$strBackupDir'),
+	);
+	
+	foreach($arrVarName as $val){
+		if(!isset($val[0])){
+			if($flag_in_html){ print("<p class=\"error\">include/config.php に ".$val[1]." が設定されていません</p>\n");}
+			else{print("Error : include/config.php - ".$val[1]." not exist\n");}
+			$flag_ok = 0;
+		}
+	}
+
+	if(!$flag_ok){ return(0); }
+
+	// execコマンドに渡す変数の汚染除去（エスケープ）
+	$strSvnCmdPath = escapeshellcmd($strSvnCmdPath);
+	$strBaseDir = escapeshellcmd($strBaseDir);
+	$strBackupDir = escapeshellcmd($strBackupDir);
+	
+	return(1);
+}
+
+
 ?>
